requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.4
python-dotenv==1.0.1
aiomqtt==1.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiofiles==23.2.1
